Updated facet_generator in bipartite_to_max_facet
The last facet of a dataset was always missing when using the facet_generator() function. The '' for '' loop in facet_generator() could not yield the last facet because when the loop is done iterating over the sorted edges, it cannot define a new ''curr_facet'' making it impossible to enter the conditional block if curr_facet != prev_facet: and yield. Simply fixed by yielding ''facet_content' after the loop.